+++ frontend/src/layouts/DefaultLayout.test.ts
@@ -1,5 +1,7 @@
 import { ApolloError } from '@apollo/client/errors'
+import { provideApolloClient } from '@vue/apollo-composable'
 import { flushPromises, mount } from '@vue/test-utils'
+import { createMockClient, createMockSubscription, IMockSubscription } from 'mock-apollo-client'
 import { navigate } from 'vike/client/router'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { Component, h } from 'vue'
@@ -9,7 +11,7 @@ import UserDropdown from '#components/menu/UserDropdown.vue'
 import { joinMyTableMutation } from '#mutations/joinMyTableMutation'
 import { AUTH } from '#src/env'
 import { useAuthStore } from '#stores/authStore'
-import { mockClient } from '#tests/mock.apolloClient'
+import { updateOpenTablesSubscription } from '#subscriptions/updateOpenTablesSubscription'
 import { authService } from '#tests/mock.authService'
 import { errorHandlerSpy } from '#tests/plugin.globalErrorHandler'
 
@@ -19,8 +21,14 @@ vi.mock('vike/client/router')
 vi.mocked(navigate).mockResolvedValue()
 
 const joinMyTableMutationMock = vi.fn()
+const updateOpenTablesSubscriptionMock: IMockSubscription = createMockSubscription()
+
+const mockClient = createMockClient()
 
 mockClient.setRequestHandler(joinMyTableMutation, joinMyTableMutationMock)
+mockClient.setRequestHandler(updateOpenTablesSubscription, () => updateOpenTablesSubscriptionMock)
+
+provideApolloClient(mockClient)
 
 describe('DefaultLayout', () => {
   const Wrapper = () => {
