+++ frontend/src/pages/table/Page.test.ts
@@ -1,11 +1,14 @@
 import { ApolloError } from '@apollo/client/errors'
+import { provideApolloClient } from '@vue/apollo-composable'
 import { flushPromises, mount } from '@vue/test-utils'
+import { createMockClient, createMockSubscription, IMockSubscription } from 'mock-apollo-client'
 import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { Component, h } from 'vue'
 import { VApp } from 'vuetify/components'
 
+import { currentUserQuery } from '#queries/currentUserQuery'
 import { joinTableQuery } from '#queries/joinTableQuery'
-import { mockClient } from '#tests/mock.apolloClient'
+import { updateOpenTablesSubscription } from '#subscriptions/updateOpenTablesSubscription'
 import { mockPageContext } from '#tests/mock.vikePageContext'
 import { errorHandlerSpy } from '#tests/plugin.globalErrorHandler'
 
@@ -14,11 +17,31 @@ import Route from './+route'
 import { title } from './+title'
 
 const joinTableQueryMock = vi.fn()
+const currentUserQueryMock = vi.fn()
+const updateOpenTablesSubscriptionMock: IMockSubscription = createMockSubscription()
 
+const mockClient = createMockClient()
+
+mockClient.setRequestHandler(
+  currentUserQuery,
+  currentUserQueryMock.mockResolvedValue({
+    data: {
+      currentUser: {
+        id: 666,
+        name: 'Current User',
+        username: 'currentUser',
+        table: null,
+      },
+    },
+  }),
+)
 mockClient.setRequestHandler(
   joinTableQuery,
   joinTableQueryMock.mockResolvedValue({ data: { joinTable: 'https://some-link-to-meeting.com' } }),
 )
+mockClient.setRequestHandler(updateOpenTablesSubscription, () => updateOpenTablesSubscriptionMock)
+
+provideApolloClient(mockClient)
 
 describe('Table Page', () => {
   const Wrapper = () => {
