+++ frontend/src/stores/tablesStore.spec.ts
@@ -1,14 +1,21 @@
 import { provideApolloClient } from '@vue/apollo-composable'
+import { createMockClient, createMockSubscription, IMockSubscription } from 'mock-apollo-client'
 import { setActivePinia, createPinia } from 'pinia'
-import { describe, it, expect, beforeEach } from 'vitest'
+import { vi, describe, it, expect, beforeEach } from 'vitest'
 
-import {
-  mockClient,
-  openTablesQueryMock,
-  updateOpenTablesSubscriptionMock,
-} from '#tests/mock.apolloClient'
+import { openTablesQuery } from '#src/graphql/queries/openTablesQuery'
+import { updateOpenTablesSubscription } from '#subscriptions/updateOpenTablesSubscription'
 
 import { useTablesStore } from './tablesStore'
+
+const updateOpenTablesSubscriptionMock: IMockSubscription = createMockSubscription()
+const mockClient = createMockClient()
+const openTablesQueryMock = vi.fn()
+mockClient.setRequestHandler(updateOpenTablesSubscription, () => updateOpenTablesSubscriptionMock)
+mockClient.setRequestHandler(
+  openTablesQuery,
+  openTablesQueryMock.mockResolvedValue({ data: { openTables: [] } }),
+)
 
 provideApolloClient(mockClient)
